producer/requirements_wh.txt
fastapi
uvicorn
pika
requests
python-dotenv
